Adds new API that allows Components to be injected in the background/foreground of each page, also forces children to be used as ParallaxSwiperPage which helps mental overhead of new page APIs coming
