+++ .eslintrc.json
@@ -24,6 +24,7 @@
     "react-native/no-color-literals": 0,
     "class-property/class-property-semicolon": 2,
     "prefer-class-properties/prefer-class-properties": 2,
-    "react/require-default-props": 0
+    "react/require-default-props": 0,
+    "max-len": ["error", 80]
   }
 }
