+++ package.json
@@ -2,7 +2,7 @@
   "name": "react-native-parallax-swiper",
   "version": "0.5.4",
   "description": "Full Screen Parallax Swiper Allowing Arbitrary UI Injection",
-  "main": "index.js",
+  "main": "src/index.js",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/zachgibson/react-native-parallax-swiper.git"
