+++ src/index.js
@@ -1,0 +1,4 @@
+import ParallaxSwiper from './ParallaxSwiper';
+import ParallaxSwiperPage from './ParallaxSwiperPage';
+
+export { ParallaxSwiper, ParallaxSwiperPage };
